requirements.txt
black==20.8b1
isort==5.7.0
Pillow==8.1.0
pre-commit==2.9.3
